Reorganized tests so that any tests requiring manually setting up OBS first are in the 'manual' package. By modifying the gradle build file, now those tests can be excluded and the rest of the integration tests can be run automatically
