+++ src/integrationTest/java/net/twasi/obsremotejava/test/AbstractObsE2ETest.java
@@ -6,67 +6,46 @@ import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.function.Consumer;
 import net.twasi.obsremotejava.OBSRemoteController;
-import org.junit.jupiter.api.BeforeAll;
 
 public abstract class AbstractObsE2ETest {
 
-  static OBSRemoteController remote;
-  BlockingQueue resultQueue = new LinkedBlockingQueue();
+  protected static OBSRemoteController remote;
+  protected BlockingQueue resultQueue = new LinkedBlockingQueue();
 
   // Scenes
-  final static String SCENE1 = "scene1";
-  final static String SCENE2 = "scene2";
-  final static String SCENE3 = "emptyscene";
+  protected final static String SCENE1 = "scene1";
+  protected final static String SCENE2 = "scene2";
+  protected final static String SCENE3 = "emptyscene";
 
   // Sources
-  final static String SOURCE_TEXT_SCENE1 = "scenename1";
-  final static String SOURCE_TEXT_SCENE2 = "scenename2";
-  final static String SOURCE_MEDIA = "media";
-  final static String SOURCE_VLC_MEDIA = "vlc-media";
-  final static String SOURCE_RED_SQUARE = "red_square";
-  final static String SOURCE_RED_SQUARE_FILTER = "Color Correction";
-  final static String SOURCE_BROWSER = "browser";
-  final static String SOURCE_GROUP = "group";
-  final static String SOURCE_GROUP_TEXT = "grouptext";
+  protected final static String SOURCE_TEXT_SCENE1 = "scenename1";
+  protected final static String SOURCE_TEXT_SCENE2 = "scenename2";
+  protected final static String SOURCE_MEDIA = "media";
+  protected final static String SOURCE_VLC_MEDIA = "vlc-media";
+  protected final static String SOURCE_RED_SQUARE = "red_square";
+  protected final static String SOURCE_RED_SQUARE_FILTER = "Color Correction";
+  protected final static String SOURCE_BROWSER = "browser";
+  protected final static String SOURCE_GROUP = "group";
+  protected final static String SOURCE_GROUP_TEXT = "grouptext";
 
   // Transitions and other OBS hard-coded names
-  final static String TRANSITION_SLIDE = "Slide";
-  final static String TRANSITION_CUT = "Cut";
-  final static String TRANSITION_FADE = "Fade";
-  final static String SOURCE_OBS_MIC = "Mic/Aux";
-  final static String SOURCE_OBS_AUDIO = "Desktop Audio";
-
-  @BeforeAll
-  static void beforeAll() {
-
-    // Connect to OBS
-    remote = new OBSRemoteController("ws://localhost:4444", false);
-    remote.registerConnectionFailedCallback(message -> {
-      fail("Failed to connect to OBS: " + message);
-    });
-    remote.registerOnError((message, throwable) -> {
-      fail("Failed to connect to OBS due to error: " + message);
-    });
-    remote.connect();
-
-    try {
-      Thread.sleep(1000);
-    } catch (InterruptedException e) {
-      e.printStackTrace();
-    }
-  }
+  protected final static String TRANSITION_SLIDE = "Slide";
+  protected final static String TRANSITION_CUT = "Cut";
+  protected final static String TRANSITION_FADE = "Fade";
+  protected final static String SOURCE_OBS_MIC = "Mic/Aux";
+  protected final static String SOURCE_OBS_AUDIO = "Desktop Audio";
 
   //  Test Helpers
-  void obsShould(String expected) {
+  protected void obsShould(String expected) {
     obsShould(expected, 3);
   }
 
-  void obsShould(String expected, int secondsTimeout) {
+  protected void obsShould(String expected, int secondsTimeout) {
     System.out.println(">>> OBS SHOULD: " + expected);
     countDownFrom(secondsTimeout);
   }
 
-  void countDownFrom(int seconds) {
+  protected void countDownFrom(int seconds) {
     for(int i = seconds; i > 0; i--) {
       System.out.println("> " + i);
       try {
@@ -77,7 +56,7 @@ public abstract class AbstractObsE2ETest {
     }
   }
 
-  static void connectToObs() {
+  protected static void connectToObs() {
     remote = new OBSRemoteController("ws://localhost:4444", false);
     remote.registerConnectionFailedCallback(message -> {
       fail("Failed to connect to OBS: " + message);
@@ -94,7 +73,7 @@ public abstract class AbstractObsE2ETest {
     }
   }
 
-  void setupObs() {
+  protected void setupObs() {
 
     // Cleanup all scenes
     cleanupScenes();
@@ -107,7 +86,7 @@ public abstract class AbstractObsE2ETest {
     });
   }
 
-  void cleanupScenes() {
+  protected void cleanupScenes() {
     // Hide all visible elements in all scenes
     remote.getScenes(sceneListResponse -> {
       sceneListResponse.getScenes().forEach(scene -> {
@@ -124,15 +103,15 @@ public abstract class AbstractObsE2ETest {
     });
   }
 
-  Consumer loggingCallback = (obj) -> {
+  protected Consumer loggingCallback = (obj) -> {
     System.out.println("Received response: " + obj);
   };
 
-  void waitReasonably() {
+  protected void waitReasonably() {
     waitReasonably(50);
   }
 
-  void waitReasonably(long ms) {
+  protected void waitReasonably(long ms) {
     try {
       Thread.sleep(ms);
     } catch (InterruptedException e) {
@@ -140,12 +119,12 @@ public abstract class AbstractObsE2ETest {
     }
   }
 
-  Consumer capturingCallback = (obj) -> {
+  protected Consumer capturingCallback = (obj) -> {
     System.out.println("Received response: " + obj + "(" + obj.getClass().getSimpleName() + ")");
     resultQueue.add(obj);
   };
 
-  <T> T getResponseAs(Class<T> clazz) {
+  protected <T> T getResponseAs(Class<T> clazz) {
     return clazz.cast(resultQueue.remove());
   }
